Refactor ITP1_7_D

+++ ITP1/ITP1_7_D/main.cpp
@@ -2,25 +2,25 @@
 #include <vector>
 using namespace std;
 
-static inline vector<vector<long>> read(vector<vector<long>> v)
+static inline vector<vector<int>> read(vector<vector<int>> v)
 {
     for (auto &x: v)
-        for (long &y : x)
+        for (int &y : x)
             cin >> y;
     return move(v);
 }
 
 int main(void)
 {
-    long n, m, l;
+    int n, m, l;
     cin >> n >> m >> l;
-    vector<vector<long>> a(n, vector<long>(m));
-    vector<vector<long>> b(m, vector<long>(l));
+    vector<vector<int>> a(n, vector<int>(m));
+    vector<vector<int>> b(m, vector<int>(l));
     a = read(move(a));
     b = read(move(b));
 
-    for (long i = 0; i < n; i++) {
-        for (long j = 0; j < l; j++) {
+    for (int i = 0; i < n; i++) {
+        for (int j = 0; j < l; j++) {
             long sum = 0;
             for (long k = 0; k < m; k++)
                 sum += a[i][k] * b[k][j];
